move packageSetEnvironment to installer
