+++ packages/dappmanager/src/calls/localProxy.ts
@@ -6,7 +6,7 @@ import {
 } from "@dappnode/hostscriptsservices";
 import { params } from "@dappnode/params";
 import { AvahiDaemonStatus, LocalProxyingStatus } from "@dappnode/common";
-import { packageSetEnvironment } from "./packageSetEnvironment.js";
+import { packageSetEnvironment } from "@dappnode/installer";
 import { listPackageNoThrow } from "@dappnode/dockerapi";
 
 /**
