+++ packages/dappmanager/src/modules/wireguard/client.ts
@@ -1,7 +1,7 @@
 import fetch from "node-fetch";
 import { urlJoin } from "@dappnode/utils";
 import { params } from "@dappnode/params";
-import { packageSetEnvironment } from "../../calls/packageSetEnvironment.js";
+import { packageSetEnvironment } from "@dappnode/installer";
 import { ComposeFileEditor } from "@dappnode/dockercompose";
 import { WireguardDeviceCredentials } from "@dappnode/common";
