+++ packages/installer/src/packages/index.ts
@@ -2,3 +2,4 @@ export { packageGet } from "./packageGet.js";
 export { packageInstall } from "./packageInstall.js";
 export { packageRemove } from "./packageRemove.js";
 export { packagesGet, sortPackages } from "./packagesGet.js";
+export { packageSetEnvironment } from "./packageSetEnvironment.js";
